move k8s base functions into k8s package
