+++ pkg/transformer/kubernetes/kubernetes.go
@@ -18,7 +18,7 @@ package kubernetes
 
 import (
 	"fmt"
-	"os"
+	"strconv"
 
 	"github.com/Sirupsen/logrus"
 	deployapi "github.com/openshift/origin/pkg/deploy/api"
@@ -26,43 +26,262 @@ import (
 	"github.com/skippbox/kompose/pkg/transformer"
 
 	"k8s.io/kubernetes/pkg/api"
+	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/apis/extensions"
+	client "k8s.io/kubernetes/pkg/client/unversioned"
 	"k8s.io/kubernetes/pkg/runtime"
+	"k8s.io/kubernetes/pkg/util/intstr"
 )
 
 type Kubernetes struct {
 }
 
-func (k *Kubernetes) Transform(komposeObject kobject.KomposeObject, opt kobject.ConvertOptions) (map[string][]byte, map[string][]byte, map[string][]byte, map[string][]byte, map[string][]byte, []string){
-	mServices := make(map[string][]byte)
-	mReplicationControllers := make(map[string][]byte)
-	mDeployments := make(map[string][]byte)
-	mDaemonSets := make(map[string][]byte)
+// Init RC object
+func InitRC(name string, service kobject.ServiceConfig, replicas int) *api.ReplicationController {
+	rc := &api.ReplicationController{
+		TypeMeta: unversioned.TypeMeta{
+			Kind:       "ReplicationController",
+			APIVersion: "v1",
+		},
+		ObjectMeta: api.ObjectMeta{
+			Name: name,
+			//Labels: map[string]string{"service": name},
+		},
+		Spec: api.ReplicationControllerSpec{
+			Selector: map[string]string{"service": name},
+			Replicas: int32(replicas),
+			Template: &api.PodTemplateSpec{
+				ObjectMeta: api.ObjectMeta{
+				//Labels: map[string]string{"service": name},
+				},
+				Spec: api.PodSpec{
+					Containers: []api.Container{
+						{
+							Name:  name,
+							Image: service.Image,
+						},
+					},
+				},
+			},
+		},
+	}
+	return rc
+}
 
+// Init SC object
+func InitSC(name string, service kobject.ServiceConfig) *api.Service {
+	sc := &api.Service{
+		TypeMeta: unversioned.TypeMeta{
+			Kind:       "Service",
+			APIVersion: "v1",
+		},
+		ObjectMeta: api.ObjectMeta{
+			Name: name,
+			//Labels: map[string]string{"service": name},
+		},
+		Spec: api.ServiceSpec{
+			Selector: map[string]string{"service": name},
+		},
+	}
+	return sc
+}
+
+// Init DC object
+func InitDC(name string, service kobject.ServiceConfig, replicas int) *extensions.Deployment {
+	dc := &extensions.Deployment{
+		TypeMeta: unversioned.TypeMeta{
+			Kind:       "Deployment",
+			APIVersion: "extensions/v1beta1",
+		},
+		ObjectMeta: api.ObjectMeta{
+			Name:   name,
+			Labels: map[string]string{"service": name},
+		},
+		Spec: extensions.DeploymentSpec{
+			Replicas: int32(replicas),
+			Selector: &unversioned.LabelSelector{
+				MatchLabels: map[string]string{"service": name},
+			},
+			//UniqueLabelKey: p.Name,
+			Template: api.PodTemplateSpec{
+				ObjectMeta: api.ObjectMeta{
+					Labels: map[string]string{"service": name},
+				},
+				Spec: api.PodSpec{
+					Containers: []api.Container{
+						{
+							Name:  name,
+							Image: service.Image,
+						},
+					},
+				},
+			},
+		},
+	}
+	return dc
+}
+
+// Init DS object
+func InitDS(name string, service kobject.ServiceConfig) *extensions.DaemonSet {
+	ds := &extensions.DaemonSet{
+		TypeMeta: unversioned.TypeMeta{
+			Kind:       "DaemonSet",
+			APIVersion: "extensions/v1beta1",
+		},
+		ObjectMeta: api.ObjectMeta{
+			Name: name,
+		},
+		Spec: extensions.DaemonSetSpec{
+			Template: api.PodTemplateSpec{
+				ObjectMeta: api.ObjectMeta{
+					Name: name,
+				},
+				Spec: api.PodSpec{
+					Containers: []api.Container{
+						{
+							Name:  name,
+							Image: service.Image,
+						},
+					},
+				},
+			},
+		},
+	}
+	return ds
+}
+
+// Configure the container ports.
+func ConfigPorts(name string, service kobject.ServiceConfig) []api.ContainerPort {
+	ports := []api.ContainerPort{}
+	for _, port := range service.Port {
+		var p api.Protocol
+		switch port.Protocol {
+		default:
+			p = api.ProtocolTCP
+		case kobject.ProtocolTCP:
+			p = api.ProtocolTCP
+		case kobject.ProtocolUDP:
+			p = api.ProtocolUDP
+		}
+		ports = append(ports, api.ContainerPort{
+			ContainerPort: port.ContainerPort,
+			Protocol:      p,
+		})
+	}
+
+	return ports
+}
+
+// Configure the container service ports.
+func ConfigServicePorts(name string, service kobject.ServiceConfig) []api.ServicePort {
+	servicePorts := []api.ServicePort{}
+	for _, port := range service.Port {
+		if port.HostPort == 0 {
+			port.HostPort = port.ContainerPort
+		}
+		var p api.Protocol
+		switch port.Protocol {
+		default:
+			p = api.ProtocolTCP
+		case kobject.ProtocolTCP:
+			p = api.ProtocolTCP
+		case kobject.ProtocolUDP:
+			p = api.ProtocolUDP
+		}
+		var targetPort intstr.IntOrString
+		targetPort.IntVal = port.ContainerPort
+		targetPort.StrVal = strconv.Itoa(int(port.ContainerPort))
+		servicePorts = append(servicePorts, api.ServicePort{
+			Name:       strconv.Itoa(int(port.HostPort)),
+			Protocol:   p,
+			Port:       port.HostPort,
+			TargetPort: targetPort,
+		})
+	}
+	return servicePorts
+}
+
+// Configure the container volumes.
+func ConfigVolumes(service kobject.ServiceConfig) ([]api.VolumeMount, []api.Volume) {
+	volumesMount := []api.VolumeMount{}
+	volumes := []api.Volume{}
+	volumeSource := api.VolumeSource{}
+	for _, volume := range service.Volumes {
+		name, host, container, mode, err := transformer.ParseVolume(volume)
+		if err != nil {
+			logrus.Warningf("Failed to configure container volume: %v", err)
+			continue
+		}
+
+		// if volume name isn't specified, set it to a random string of 20 chars
+		if len(name) == 0 {
+			name = transformer.RandStringBytes(20)
+		}
+		// check if ro/rw mode is defined, default rw
+		readonly := len(mode) > 0 && mode == "ro"
+
+		volumesMount = append(volumesMount, api.VolumeMount{Name: name, ReadOnly: readonly, MountPath: container})
+
+		if len(host) > 0 {
+			volumeSource = api.VolumeSource{HostPath: &api.HostPathVolumeSource{Path: host}}
+		} else {
+			volumeSource = api.VolumeSource{EmptyDir: &api.EmptyDirVolumeSource{}}
+		}
+
+		volumes = append(volumes, api.Volume{Name: name, VolumeSource: volumeSource})
+	}
+	return volumesMount, volumes
+}
+
+// Configure the environment variables.
+func ConfigEnvs(name string, service kobject.ServiceConfig) []api.EnvVar {
+	envs := []api.EnvVar{}
+	for _, v := range service.Environment {
+		envs = append(envs, api.EnvVar{
+			Name:  v.Name,
+			Value: v.Value,
+		})
+	}
+
+	return envs
+}
+
+func (k *Kubernetes) Transform(komposeObject kobject.KomposeObject, opt kobject.ConvertOptions) []runtime.Object {
 	var svcnames []string
 
+	// this will hold all the converted data
+	var allobjects []runtime.Object
+
 	for name, service := range komposeObject.ServiceConfigs {
+		var objects []runtime.Object
 		svcnames = append(svcnames, name)
 
-		rc := transformer.InitRC(name, service, opt.Replicas)
-		sc := transformer.InitSC(name, service)
-		dc := transformer.InitDC(name, service, opt.Replicas)
-		ds := transformer.InitDS(name, service)
+		sc := InitSC(name, service)
+
+		if opt.CreateD {
+			objects = append(objects, InitDC(name, service, opt.Replicas))
+		}
+		if opt.CreateDS {
+			objects = append(objects, InitDS(name, service))
+		}
+		if opt.CreateRC {
+			objects = append(objects, InitRC(name, service, opt.Replicas))
+		}
 
 		// Configure the environment variables.
-		envs := transformer.ConfigEnvs(name, service)
+		envs := ConfigEnvs(name, service)
 
 		// Configure the container command.
 		cmds := transformer.ConfigCommands(service)
 
 		// Configure the container volumes.
-		volumesMount, volumes := transformer.ConfigVolumes(service)
+		volumesMount, volumes := ConfigVolumes(service)
 
 		// Configure the container ports.
-		ports := transformer.ConfigPorts(name, service)
+		ports := ConfigPorts(name, service)
 
 		// Configure the service ports.
-		servicePorts := transformer.ConfigServicePorts(name, service)
+		servicePorts := ConfigServicePorts(name, service)
 		sc.Spec.Ports = servicePorts
 
 		// Configure label
@@ -107,48 +326,21 @@ func (k *Kubernetes) Transform(komposeObject kobject.KomposeObject, opt kobject.
 			meta.Annotations = annotations
 		}
 
-		// Update each supported controllers
-		UpdateController(rc, fillTemplate, fillObjectMeta)
-		UpdateController(dc, fillTemplate, fillObjectMeta)
-		UpdateController(ds, fillTemplate, fillObjectMeta)
-
-		// convert datarc to json / yaml
-		datarc, err := transformer.TransformData(rc, opt.GenerateYaml)
-		if err != nil {
-			logrus.Fatalf(err.Error())
+		// update supported controller
+		for _, obj := range objects {
+			UpdateController(obj, fillTemplate, fillObjectMeta)
 		}
 
-		// convert datadc to json / yaml
-		datadc, err := transformer.TransformData(dc, opt.GenerateYaml)
-		if err != nil {
-			logrus.Fatalf(err.Error())
-		}
-
-		// convert datads to json / yaml
-		datads, err := transformer.TransformData(ds, opt.GenerateYaml)
-		if err != nil {
-			logrus.Fatalf(err.Error())
-		}
-
-		var datasvc []byte
 		// If ports not provided in configuration we will not make service
 		if len(ports) == 0 {
 			logrus.Warningf("[%s] Service cannot be created because of missing port.", name)
-		} else if len(ports) != 0 {
-			// convert datasvc to json / yaml
-			datasvc, err = transformer.TransformData(sc, opt.GenerateYaml)
-			if err != nil {
-				logrus.Fatalf(err.Error())
-			}
+		} else {
+			objects = append(objects, sc)
 		}
-
-		mServices[name] = datasvc
-		mReplicationControllers[name] = datarc
-		mDeployments[name] = datadc
-		mDaemonSets[name] = datads
+		allobjects = append(allobjects, objects...)
 	}
 
-	return mServices, mDeployments, mDaemonSets, mReplicationControllers, nil, svcnames
+	return allobjects
 }
 
 // updateController updates the given object with the given pod template update function and ObjectMeta update function
@@ -163,9 +355,6 @@ func UpdateController(obj runtime.Object, updateTemplate func(*api.PodTemplateSp
 	case *extensions.Deployment:
 		updateTemplate(&t.Spec.Template)
 		updateMeta(&t.ObjectMeta)
-	case *extensions.ReplicaSet:
-		updateTemplate(&t.Spec.Template)
-		updateMeta(&t.ObjectMeta)
 	case *extensions.DaemonSet:
 		updateTemplate(&t.Spec.Template)
 		updateMeta(&t.ObjectMeta)
@@ -173,4 +362,24 @@ func UpdateController(obj runtime.Object, updateTemplate func(*api.PodTemplateSp
 		updateTemplate(t.Spec.Template)
 		updateMeta(&t.ObjectMeta)
 	}
+}
+
+func CreateObjects(client *client.Client, objects []runtime.Object) {
+	for _, v := range objects {
+		switch t := v.(type) {
+		case *extensions.Deployment:
+			_, err := client.Deployments(api.NamespaceDefault).Create(t)
+			if err != nil {
+				logrus.Fatalf("Error: '%v' while creating deployment: %s", err, t.Name)
+			}
+			logrus.Infof("Successfully created deployment: %s", t.Name)
+		case *api.Service:
+			_, err := client.Services(api.NamespaceDefault).Create(t)
+			if err != nil {
+				logrus.Fatalf("Error: '%v' while creating service: %s", err, t.Name)
+			}
+			logrus.Infof("Successfully created service: %s", t.Name)
+		}
+	}
+	fmt.Println("\nApplication has been deployed to Kubernetes. You can run 'kubectl get deployment,svc' for details.")
 }
